Enchancment: added entity id to component
